# perf(model): improve image hash parsing