--- a/twilight-model/benches/image_hash.rs
+++ b/twilight-model/benches/image_hash.rs
@@ -1,4 +1,4 @@
-use criterion::{criterion_group, criterion_main, Criterion};
+use criterion::{criterion_group, criterion_main, BenchmarkId, Criterion};
 use twilight_model::util::image_hash::ImageHash;
 
 fn parse_hash(hash: &str) {
@@ -6,12 +6,23 @@ fn parse_hash(hash: &str) {
 }
 
 fn criterion_benchmark(c: &mut Criterion) {
-    c.bench_function("parse non-animated hex", |b| {
-        b.iter(|| parse_hash("e15db5d5ccaa4714a6fac0514d3d2037"));
-    });
-    c.bench_function("parse animated hex", |b| {
-        b.iter(|| parse_hash("a_e15db5d5ccaa4714a6fac0514d3d2037"));
-    });
+    const ANIMATED: &str = "a_e15db5d5ccaa4714a6fac0514d3d2037";
+    const NON_ANIMATED: &str = "e15db5d5ccaa4714a6fac0514d3d2037";
+
+    c.bench_with_input(
+        BenchmarkId::new("parse non-animated hex", NON_ANIMATED),
+        &NON_ANIMATED,
+        |bencher, input| {
+            bencher.iter(|| parse_hash(input));
+        },
+    );
+    c.bench_with_input(
+        BenchmarkId::new("parse animated hex", ANIMATED),
+        &ANIMATED,
+        |bencher, input| {
+            bencher.iter(|| parse_hash(input));
+        },
+    );
 }
 
 criterion_group!(benches, criterion_benchmark);