Refactor the `ImageHash::parse` method to improve performance. This is primarily done by working directly with mutating an array of bytes instead of working with a 128 bit integer, shifting its bytes around, and then converting it into a 16-byte array, as well as avoiding some branches by using a `while let` over a `loop`. Image hashes are largely used across the model crate with 59 uses, so this results in a good cumulative win.

We see a 41% performance improvement with these changes:

```
parse non-animated hex  time:   [34.367 ns 34.767 ns 35.437 ns]
                        change: [-42.665% -41.460% -40.302%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 11 outliers among 100 measurements (11.00%)
  5 (5.00%) high mild
  6 (6.00%) high severe

parse animated hex      time:   [34.351 ns 34.366 ns 34.385 ns]
                        change: [-38.538% -38.483% -38.420%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 10 outliers among 100 measurements (10.00%)
  5 (5.00%) high mild
  5 (5.00%) high severe
```

**Todo**:

- [x] Determine if we can reasonably de-duplicate these `match` statements since `const fn`s don't support `?` -- Not feasible.