@zeylahellyer 2023-03-11
Here are the benchmarks using `Criterion::bench_with_input`:

```
Benchmarking parse non-animated hex/e15db5d5ccaa4714a6fac0514d3d2037: Collecting 100 samples in estima
parse non-animated hex/e15db5d5ccaa4714a6fac0514d3d2037
                        time:   [21.360 ns 21.371 ns 21.382 ns]
Found 20 outliers among 100 measurements (20.00%)
  7 (7.00%) low severe
  2 (2.00%) low mild
  7 (7.00%) high mild
  4 (4.00%) high severe

Benchmarking parse animated hex/a_e15db5d5ccaa4714a6fac0514d3d2037: Collecting 100 samples in estimate
parse animated hex/a_e15db5d5ccaa4714a6fac0514d3d2037
                        time:   [21.312 ns 21.337 ns 21.358 ns]
Found 29 outliers among 100 measurements (29.00%)
  16 (16.00%) low severe
  2 (2.00%) low mild
  2 (2.00%) high mild
  9 (9.00%) high severe
```